+++ src/components/modals/InstantWithdrawalModal.tsx
@@ -80,6 +80,10 @@ const InstantWithdrawalModal = ({
           queryKey: ['balances'],
           refetchType: 'all',
         });
+        queryClient.invalidateQueries({
+          queryKey: ['delegateStakes'],
+          refetchType: 'all',
+        });
 
         setShowSuccessModal(true);
       } catch (e: any) {